Added .bat file to copy assets' files

+++ simpleGradius/src/Game/Setup/Game.cpp
@@ -20,8 +20,8 @@ namespace Juego
 {
 	int gameScreen = Menu;
 
-	int screenWidth;
-	int screenHeight;
+	float screenWidth;
+	float screenHeight;
 
 	Buttons mouse;
 	int buttonOption = 0;
